+++ packages/form-render/src/widgets/antd/date.js
@@ -1,18 +1,13 @@
 import React from 'react';
 import moment from 'moment';
-import { DatePicker, TimePicker } from 'antd';
+import { DatePicker } from 'antd';
 import { getFormat } from '../../utils';
 
 // TODO: 不要使用moment，使用dayjs
-export default p => {
-  let { format = 'dateTime' } = p.schema;
-  if (p.options && p.options.format) {
-    format = p.options.format;
-  }
-  const DateComponent = format === 'time' ? TimePicker : DatePicker;
+export default ({ onChange, format, value, style, ...rest }) => {
   const dateFormat = getFormat(format);
   // week的时候会返回 2020-31周 quarter会返回 2020-Q2 需要处理之后才能被 moment
-  let _value = p.value || '';
+  let _value = value || undefined;
   if (typeof _value === 'string') {
     if (format === 'week') {
       _value = _value.substring(0, _value.length - 1);
@@ -25,17 +20,14 @@ export default p => {
     _value = moment(_value, dateFormat);
   }
 
-  const placeholderObj = p.description ? { placeholder: p.description } : {};
+  const handleChange = (value, string) => {
+    onChange(string);
+  };
 
-  const onChange = (value, string) => p.onChange(string);
-
-  const dateParams = {
-    ...placeholderObj,
-    ...p.options,
+  let dateParams = {
     value: _value,
-    style: { width: '100%' },
-    disabled: p.disabled || p.readOnly,
-    onChange,
+    style: { width: '100%', ...style },
+    onChange: handleChange,
   };
 
   // TODO: format是在options里自定义的情况，是否要判断一下要不要showTime
@@ -47,5 +39,7 @@ export default p => {
     dateParams.picker = format;
   }
 
-  return <DateComponent {...dateParams} />;
+  dateParams = { ...dateParams, ...rest };
+
+  return <DatePicker {...dateParams} />;
 };