+++ packages/form-render/CHANGELOG.md
@@ -1,5 +1,25 @@
 # Changelog
 
+### 1.0.5
+
+- [+] 新增 `validateMessages` 字段，用于覆盖默认的校验信息，详见[文档](https://x-render.gitee.io/form-render#validatemessages) ([#306](https://github.com/alibaba/x-render/issues/306))
+- [!] rules 字段无法生效的问题 ([#305](https://github.com/alibaba/x-render/issues/305))
+- [!] 修复了下拉多选框在 value = null 时会展示一个空标签的问题
+- [!] 说明（description）的 tooltip 展示气泡位置确保对齐
+
+### 1.0.4
+
+- [+] 新增时间区间组件 timeRange。通过`{type: 'range', format: 'time'}` 渲染
+- [!] 完善 ts 声明文件。([#302](https://github.com/alibaba/x-render/pull/302))
+- [!] fix 了 rules 校验和 image 校验的冲突
+- [!] 修复了 tooltip、checkbox 等样式的问题
+
+### 1.0.3
+
+- [!] fix 列表的 bind 无效的问题
+- [!] fix format: image 未校验的问题
+- [!] 默认列表样式微调
+
 ### 1.0.2
 
 - [+] 新增默认列表展示 card list。widget: list0